refactor(semantic): keep authoring lowering within quality budget
